Prevented KeyError on signal restoration
If the original_handlers dictionary is not yet populated before an error
is encountered in the main try section of await_pg_notifications(), the
"finally" section may encounter a KeyError when attempting to cleanup. I
encountered this by an early failure in the function due to some illegal
asynchonous behavior.